refactor: useLoadingAnimation
