+++ web/app/components/Sigma/NextPageButton.tsx
@@ -1,3 +1,4 @@
+import { useLoadingAnimation } from '@/app/lib/hooks/useLoadingAnimation'
 import { Icon } from '@tendenz/icons'
 import { motion } from 'framer-motion'
 import { ButtonHTMLAttributes, DetailedHTMLProps } from 'react'
@@ -10,6 +11,8 @@ export function NextPageButton({
 	ButtonHTMLAttributes<HTMLButtonElement>,
 	HTMLButtonElement
 >) {
+	const { loadingAnimation } = useLoadingAnimation()
+
 	return (
 		<motion.div
 			whileInView={{
@@ -22,7 +25,12 @@ export function NextPageButton({
 			}}
 			className="my-10 flex items-center justify-center"
 		>
-			<FreeFloatButton {...props} aria-label="load more" className="text-2xl">
+			<FreeFloatButton
+				{...props}
+				aria-label="load more"
+				className="text-2xl"
+				disabled={loadingAnimation}
+			>
 				<Icon name="radix-icons/chevron-down" />
 			</FreeFloatButton>
 		</motion.div>
