+++ web/app/page.tsx
@@ -1,7 +1,7 @@
 import { ComingSoon } from './components/comingsoon/ComingSoon'
 import { FilterCluster } from './components/filtercluster/FilterCluster'
 import Hero from './components/hero/Hero'
-import { SigmaAccordion } from './components/sigma/SigmaAccordion'
+import { SigmaTable } from './components/sigma/SigmaTable'
 import UsStocksHeader from './components/usstocksheader/UsStocksHeader'
 import {
 	DEFAULT_MARKET_CAP,
@@ -40,7 +40,7 @@ export default async function Home() {
 				<SWRConfigProvider fallback={fallback}>
 					<UsStocksHeader />
 					<FilterCluster />
-					<SigmaAccordion />
+					<SigmaTable />
 				</SWRConfigProvider>
 			</section>
